Added Staff as a member variable and checking in GoBack method if Student or Staff dashboard
