+++ MySupervisn-Team1/EditProfileStudent.xaml.cs
@@ -12,7 +12,6 @@ namespace MySupervisn_Team1
     public partial class EditProfileStudent : Window
     {
         private Student mStudent;
-        private Staff mStaff;
         Student temp; // ...
 
 
@@ -22,28 +21,6 @@ namespace MySupervisn_Team1
 
             mStudent = pStudent;
             temp = student;
-        }
-        public EditProfileStudent(Staff pStaff)
-        {
-            InitializeComponent();
-
-            mStaff = pStaff;
-        }   
-
-        private void GoBack_Click(object sender, RoutedEventArgs e)
-        {
-            // test
-            Close();
-            if (mStudent != null)
-            {
-                StudentDashboard studentDashboardWindow = new StudentDashboard(mStudent);
-                studentDashboardWindow.Show();
-            }
-            else
-            {
-                StaffDashboard staffDashboardWindow = new StaffDashboard(mStaff);
-                staffDashboardWindow.Show();
-            }
         }
             
         private void BtnChangeName_Click(object sender, RoutedEventArgs e)
@@ -81,7 +58,14 @@ namespace MySupervisn_Team1
             {
                 MessageBox.Show(ex.ToString());
             }
+        }
 
+        private void GoBack_Click(object sender, RoutedEventArgs e)
+        {
+            Close();
+
+            StudentDashboard studentDashboardWindow = new StudentDashboard(mStudent);
+            studentDashboardWindow.Show();
         }
 
     }
