--- MySupervisn-Team1/LectureCheck-in.xaml.cs
@@ -1,37 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
-
-namespace MySupervisn_Team1
-{
-    /// <summary>
-    /// Interaction logic for Window4.xaml
-    /// </summary>
-    public partial class Window4 : Window
-    {
-        public Window4()
-        {
-            InitializeComponent();
-        }
-
-        private void Check_in_Click(object sender, RoutedEventArgs e)
-        {
-
-        }
-
-        private void Go_Back_Click(object sender, RoutedEventArgs e)
-        {
-
-        }
-    }
-}
