+++ MySupervisn-Team1/LoginWindow.xaml.cs
@@ -137,7 +137,7 @@ namespace MySupervisn_Team1
             else {
               
                    MessageBox.Show("Invalid credentials", "Access Denied", MessageBoxButton.OK, MessageBoxImage.Warning);
-                InitializeComponent();
+                   InitializeComponent();
             
             }
         }
