+++ MySupervisn-Team1/MySupervisn-Team1.csproj
@@ -140,9 +140,6 @@
     <Compile Include="StudentSearch.xaml.cs">
       <DependentUpon>StudentSearch.xaml</DependentUpon>
     </Compile>
-    <Compile Include="LectureCheck-in.xaml.cs">
-      <DependentUpon>LectureCheck-in.xaml</DependentUpon>
-    </Compile>
     <Page Include="AddDELETEUserWindow.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
@@ -195,10 +192,6 @@
       <Generator>MSBuild:Compile</Generator>
     </Page>
     <Page Include="StudentSearch.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="LectureCheck-in.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -256,8 +249,6 @@
   <ItemGroup>
     <Folder Include="Windows\" />
   </ItemGroup>
-<<<<<<< HEAD
-=======
   <ItemGroup>
     <Content Include="DataBase\Users.mdf">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
@@ -267,7 +258,6 @@
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
   </ItemGroup>
->>>>>>> master
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
     <PropertyGroup>
